Fixed AlphaNumComparer on long digits

+++ FileScout/AlphaNumComparer.cs
@@ -35,8 +35,8 @@ namespace FileScout
                 return 0;
             }
 
-            int thisMarker = 0, thisNumericChunk = 0;
-            int thatMarker = 0, thatNumericChunk = 0;
+            long thisMarker = 0, thisNumericChunk = 0;
+            long thatMarker = 0, thatNumericChunk = 0;
 
             while ((thisMarker < s1.Length) || (thatMarker < s2.Length))
             {
@@ -48,8 +48,8 @@ namespace FileScout
                 {
                     return 1;
                 }
-                char thisCh = s1[thisMarker];
-                char thatCh = s2[thatMarker];
+                char thisCh = s1[(int)thisMarker];
+                char thatCh = s2[(int)thatMarker];
 
                 StringBuilder thisChunk = new StringBuilder();
                 StringBuilder thatChunk = new StringBuilder();
@@ -61,7 +61,7 @@ namespace FileScout
 
                     if (thisMarker < s1.Length)
                     {
-                        thisCh = s1[thisMarker];
+                        thisCh = s1[(int)thisMarker];
                     }
                 }
 
@@ -72,7 +72,7 @@ namespace FileScout
 
                     if (thatMarker < s2.Length)
                     {
-                        thatCh = s2[thatMarker];
+                        thatCh = s2[(int)thatMarker];
                     }
                 }
 
@@ -80,8 +80,8 @@ namespace FileScout
                 // If both chunks contain numeric characters, sort them numerically
                 if (char.IsDigit( thisChunk[0] ) && char.IsDigit( thatChunk[0] ))
                 {
-                    thisNumericChunk = Convert.ToInt32( thisChunk.ToString() );
-                    thatNumericChunk = Convert.ToInt32( thatChunk.ToString() );
+                    thisNumericChunk = Convert.ToInt64( thisChunk.ToString() );
+                    thatNumericChunk = Convert.ToInt64( thatChunk.ToString() );
 
                     if (thisNumericChunk < thatNumericChunk)
                     {
